Invoice Extract V4/requirements.txt
pyMuPDF
openai==1.54.4
pandas==1.5.3
Pillow==10.4.0
pydantic==2.9.2
pytesseract==0.3.13
streamlit==1.31.0
python-dotenv==1.0.1
